Consolidate drop indicator logic
